requirements.txt
pyserial>=3.5
supabase>=2.0.0
python-dotenv>=1.0.0
pyinstaller>=5.0.0